Split utils fly code into its own file
